lots of stuff

+++ src/main/java/sk/krizan/fitness_app_be/service/impl/ProfileServiceImpl.java
@@ -2,11 +2,13 @@ package sk.krizan.fitness_app_be.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import sk.krizan.fitness_app_be.controller.exception.ForbiddenException;
 import sk.krizan.fitness_app_be.controller.exception.IllegalOperationException;
 import sk.krizan.fitness_app_be.controller.exception.NotFoundException;
 import sk.krizan.fitness_app_be.controller.request.CreateProfileRequest;
 import sk.krizan.fitness_app_be.model.entity.Profile;
 import sk.krizan.fitness_app_be.model.entity.User;
+import sk.krizan.fitness_app_be.model.enums.Role;
 import sk.krizan.fitness_app_be.model.mapper.ProfileMapper;
 import sk.krizan.fitness_app_be.repository.ProfileRepository;
 import sk.krizan.fitness_app_be.service.api.ProfileService;
@@ -34,19 +36,28 @@ public class ProfileServiceImpl implements ProfileService {
     }
 
     @Override
-    public Profile createProfile(CreateProfileRequest request) {
-        User user = userService.getUserById(request.userId());
+    public Profile createProfile(CreateProfileRequest request, Long userId) {
+        User user = userService.getUserById(userId);
         if (user.getProfile() != null) {
             throw new IllegalOperationException(
                 "User { " + user.getEmail() + " } already has an assigned profile.");
         }
 
-        return ProfileMapper.createProfileRequestToProfile(request, user);
+        Profile profile = ProfileMapper.createProfileRequestToProfile(request, user);
+        user.setProfile(profile);
+
+        return profileRepository.save(profile);
     }
 
     @Override
     public Long deleteProfile(Long id) {
+        User currentUser = userService.getCurrentUser();
         Profile profile = getProfileById(id);
+
+        if (profile.getUser() != currentUser && !currentUser.getRoles().contains(Role.ADMIN)) {
+            throw new ForbiddenException();
+        }
+
         profileRepository.delete(profile);
         return profile.getId();
     }
